backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
websockets==13.1
httpx==0.28.1
pydantic-settings==2.7.1
email-validator==2.2.0
jinja2==3.1.4

# GIS — 산림청 SHP 적재·교차연산용 (PostGIS 없이 shapely 로 Python 측 계산)
shapely==2.0.6
pyogrio==0.10.0
pyproj==3.7.0
geopandas==1.0.1  # pyogrio.read_dataframe 이 내부적으로 요구
# 산사태위험 래스터(GeoTIFF) zonal stats 용 — 시스템 libexpat 필요
rasterio==1.3.11

